wip: vlidation tools
