+++ examples/latergram/src/lib/agent/code_agent.ts
@@ -1,5 +1,5 @@
 import { createOpenRouter } from '@openrouter/ai-sdk-provider';
-import { validatedTonkTools as tonkTools } from './tools/validated-tonk-tools';
+import { tonkTools } from './tools/tonk-tools';
 
 // Get OpenRouter API key from environment
 const getOpenRouterApiKey = (): string => {
