+++ examples/latergram/src/lib/agent/index.ts
@@ -4,4 +4,4 @@ export { getAgentService, resetAgentService } from './agent-service';
 export type { AgentServiceOptions } from './agent-service';
 export { getChatHistory, resetChatHistory } from './chat-history';
 export type { ChatMessage, ChatHistoryData } from './chat-history';
-export { tonkTools } from './tools/tonk-tools';
+export { tonkTools } from './tools/tonk-tools.ts';
